+++ src/main/java/net/silentchaos512/gems/api/chaos/ChaosEvent.java
@@ -12,15 +12,21 @@ public class ChaosEvent {
     private final int minChaos;
     private final int maxChaos;
     private final int chaosDissipated;
+    private final String configComment;
     private final BiFunction<PlayerEntity, Integer, Boolean> action;
 
-    public ChaosEvent(float chance, int cooldownTimeInSeconds, int minChaos, int maxChaos, int chaosDissipated, BiFunction<PlayerEntity, Integer, Boolean> action) {
+    public ChaosEvent(float chance, int cooldownTimeInSeconds, int minChaos, int maxChaos, int chaosDissipated, String configComment, BiFunction<PlayerEntity, Integer, Boolean> action) {
         this.chance = chance;
         this.cooldownTime = cooldownTimeInSeconds;
         this.minChaos = minChaos;
         this.maxChaos = maxChaos;
         this.chaosDissipated = chaosDissipated;
+        this.configComment = configComment;
         this.action = action;
+    }
+
+    public String getConfigComment() {
+        return configComment;
     }
 
     public boolean tryActivate(PlayerEntity entity, int chaos) {